--- frontend/src/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton.tsx
@@ -1,7 +1,0 @@
-import { StyledContainer } from '@/pages/GroupInitPage/components/GroupCreateFormSubmitButton/GroupCreateFormSubmitButton.styles';
-
-function GroupCreateFormSubmitButton() {
-  return <StyledContainer type="submit">생성하기</StyledContainer>;
-}
-
-export default GroupCreateFormSubmitButton;
