+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsCreateForm/GroupInitFormsCreateForm.styles.tsx
