+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsCreateFormSubmitButton/GroupInitFormsCreateFormSubmitButton.styles.tsx
