+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsCreateFormSubmitButton/GroupInitFormsCreateFormSubmitButton.tsx
@@ -1,0 +1,7 @@
+import { StyledContainer } from '@/pages/GroupInitPage/components/GroupInitFormsCreateFormSubmitButton/GroupInitFormsCreateFormSubmitButton.styles';
+
+function GroupInitFormsCreateFormSubmitButton() {
+  return <StyledContainer type="submit">생성하기</StyledContainer>;
+}
+
+export default GroupInitFormsCreateFormSubmitButton;
