+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsParticipateForm/GroupInitFormsParticipateForm.styles.tsx
