+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsParticipateFormInvitationCodeInput/GroupInitFormsParticipateFormInvitationCodeInput.styles.tsx
