+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsParticipateFormSubmitButton/GroupInitFormsParticipateFormSubmitButton.styles.tsx
