+++ frontend/src/pages/GroupInitPage/components/GroupInitFormsParticipateFormSubmitButton/GroupInitFormsParticipateFormSubmitButton.tsx
@@ -1,0 +1,7 @@
+import { StyledContainer } from '@/pages/GroupInitPage/components/GroupInitFormsParticipateFormSubmitButton/GroupInitFormsParticipateFormSubmitButton.styles';
+
+function GroupInitFormsParticipateFormSubmitButton() {
+  return <StyledContainer type="submit">참가하기</StyledContainer>;
+}
+
+export default GroupInitFormsParticipateFormSubmitButton;
