--- frontend/src/pages/GroupInitPage/components/GroupParticipateFormSubmitButton/GroupParticipateFormSubmitButton.tsx
@@ -1,7 +1,0 @@
-import { StyledContainer } from '@/pages/GroupInitPage/components/GroupParticipateFormSubmitButton/GroupParticipateFormSubmitButton.styles';
-
-function GroupParticipateFormSubmitButton() {
-  return <StyledContainer type="submit">참가하기</StyledContainer>;
-}
-
-export default GroupParticipateFormSubmitButton;
